Get rid of PrettyInfo
The problem with PrettyInfo is that since it is a regular function, we always
call it, always evaluate its arguments, and always assemble the InfoString --
even if it then turns out that the given info level is not active.

In contrast, GAP tries hard to induce as little overhead as possible for Info
statements, e.g. its arguments are not evaluated unless necessary.