Improve FetchBrentFactors
This function now gets two arguments: an URL of a file containing
factors and true/false for updating/not updating the local factor files.

The comment before the function explains how to update the data of
factors of numbers of form a^n - 1. Currently, there seems much activity
to extend this data collection.

In particular, these data are no longer maintained by Richard Brent
but by Jonathan Crombie.